# openapi.yaml
openapi: 3.0.3
info:
  title: Ester API
  version: "2.0.0"
  description: >
    Obnovlennyy OpenAPI dlya «Ester v2.0».
    Dobavleny marshruty:
      - /mem/kg/* — graf znaniy (uzly/rebra, eksport/import)
      - /mem/hypothesis — khranilische gipotez (list/add/feedback)
      - Aliasy pamyati pod /mem/* (flashback/alias/compact)
      - /dreams/* — zapusk «Snov»
      - /events/* — publikatsiya/lenta sobytiy; /tasks/* — planirovschik
      - /ingest/* — OCR/ASR/Image Caption/Code Scan/Dedup/Bulk
      - /ops/* — GDPR eksport/udalenie personalnykh dannykh
servers:
  - url: /
    description: Relative server (behind reverse proxy)
...
security:
  - bearerAuth: []
paths:
...
  /dreams/run:
    post:
      tags: [Dreams]
      summary: Ruchnoy zapusk «Snov»
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                rules:
                  type: array
                  items:
                    type: object
                    properties:
                      query: { type: string, default: "*" }
                      k: { type: integer, default: 200 }
                      ngram: { type: integer, default: 3 }
                      min_cluster_size: { type: integer, default: 2 }
                      max_hypotheses_per_cluster: { type: integer, default: 2 }
                      topic_hint: { type: string }
      responses:
        "200": { description: Otchet, content: { application/json: { schema: { type: object }}}}
        "401": { $ref: "#/components/responses/Unauthorized" }
  /dreams/last_report:
    get:
      tags: [Dreams]
      summary: Poluchit posledniy otchet po «Snam»
      responses:
        "200": { description: OK }
        "401": { $ref: "#/components/responses/Unauthorized" }

  # -------- EVENTS & TASKS
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Trebuetsya avtorizatsiya (Bearer JWT)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      properties:
        ok: { type: boolean, example: false }
        error: { type: string, example: "text required" }
...